Download management
Checks HTTP status on HEAD
Gives up when the URL gives us garbage
Split download and unexport functionality into another source file
